# doc/openapi/spec.yml
openapi: 3.0.3
info:
  title: Quiz API
  license:
    name: Michael COLL
    url: https://github.com/michaelcoll
  description: |
    Quiz App backend
  version: v1
servers:
- url: https://localhost:8080/api/v1
  description: Local Backend
...
paths:
...
  /quiz/{sha1}/class/{classId}:
    post:
      tags:
      - quiz
      summary: v1/quiz/{sha1}/class/{classId}
      description: Grant visibility to the given class
      operationId: createQuizClassVisibility
      parameters:
      - name: sha1
        in: path
        description: The sha1 of a quiz
        required: true
        schema:
          type: string
          nullable: false
          example: 'c152b2d0a2509a82ea5e8a6ae22fea55c7221002'
      - name: classId
        in: path
        description: The id of a class
        required: true
        schema:
          type: string
          format: uuid
          nullable: false
          example: "f6567dd8-e069-418e-8893-7d22fcf12459"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: "the class can access the quiz"
        "400":
          description: Quiz or class were not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
      security:
      - github: [ ]
    delete:
      tags:
      - quiz
      summary: v1/quiz/{sha1}/class/{classId}
      description: Remove visibility to the given class
      operationId: deleteQuizClassVisibility
      parameters:
      - name: sha1
        in: path
        description: The sha1 of a quiz
        required: true
        schema:
          type: string
          nullable: false
          example: 'c152b2d0a2509a82ea5e8a6ae22fea55c7221002'
      - name: classId
        in: path
        description: The id of a class
        required: true
        schema:
          type: string
          format: uuid
          nullable: false
          example: "f6567dd8-e069-418e-8893-7d22fcf12459"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: "the class can no longer access the quiz"
        "400":
          description: Quiz or class were not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
      security:
      - github: [ ]
...
  /class:
    get:
      tags:
      - class
      summary: v1/class
      description: 'The list of the classes <br /> ⚠️ Required role : **TEACHER**'
      operationId: classList
      parameters:
      - name: Range
        in: header
        description: The range header that specify the element to return
        schema:
          type: string
          nullable: false
          example: 'class=0-25'
      responses:
        "200":
          description: Success
          headers:
            Content-Range:
              description: The header specifying how much elements are return and in total
              schema:
                type: string
                example: "class 0-2/2"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Class'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
      security:
      - github: [ ]
    post:
      tags:
      - class
      summary: v1/class
      description: 'Add a classes <br /> ⚠️ Required role : **TEACHER**'
      operationId: classCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassRequestBody'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: "class created"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
      security:
      - github: [ ]
  /class/{classId}:
    put:
      tags:
      - class
      summary: v1/class/{classId}
      description: 'Update a classes <br /> ⚠️ Required role : **TEACHER**'
      operationId: classUpdate
      parameters:
      - name: classId
        in: path
        description: The id of the class
        required: true
        schema:
          type: string
          format: uuid
          nullable: false
          example: '497f6eca-6276-4993-bfeb-53cbbbba6f08'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassRequestBody'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: "class updated"
        "400":
          description: the class id is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: "invalid classId"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
      security:
      - github: [ ]
    delete:
      tags:
      - class
      summary: v1/class/{classId}
      description: 'Delete a classes <br /> ⚠️ Required role : **ADMIN**'
      operationId: classDelete
      parameters:
      - name: classId
        in: path
        description: The id of the class
        required: true
        schema:
          type: string
          format: uuid
          nullable: false
          example: '497f6eca-6276-4993-bfeb-53cbbbba6f08'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: "class deleted"
        "400":
          description: the class id is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: "invalid classId"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
      security:
      - github: [ ]
components:
  securitySchemes:
    github:
      type: oauth2
      name: github
      scheme: bearer
      bearerFormat: jwt
      flows:
        authorizationCode:
          scopes: {
                    'user'
          }
          authorizationUrl: https://github.com/login/oauth/authorize
          tokenUrl: https://github.com/login/oauth/access_token
  schemas:
...
    Message:
      type: object
      properties:
        message:
          type: string
          description: The message
          nullable: false
          example: 'a message'
...
    Class:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The id of the class
          nullable: false
        name:
          type: string
          description: The name of the class
          nullable: false
          example: 'Promotion 2023-2024'
    ClassRequestBody:
      type: object
      properties:
        name:
          type: string
          description: The name of the class
          nullable: false
          example: 'Promotion 2023-2024'