# internal/web/api/model/class-request-body.ts
/**
 * Quiz API
 * Quiz App backend
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface ClassRequestBody
 */
export interface ClassRequestBody {
  /**
   * The name of the class
   * @type {string}
   * @memberof ClassRequestBody
   */
  name?: string;
}