# internal/web/api/model/class.ts
/**
 * Quiz API
 * Quiz App backend
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface Class
 */
export interface Class {
  /**
   * The id of the class
   * @type {string}
   * @memberof Class
   */
  id?: string;
  /**
   * The name of the class
   * @type {string}
   * @memberof Class
   */
  name?: string;
}